// src/app/shared/interfaces/query.interface.ts
import {ISearch} from './search.interface';
import {HttpParams} from '@angular/common/http';

export interface IQuery {
  page?: number;
  limit?: number;
  skip?: number;
  sort?: string;
  order?: string;
  range?: string;
  filed?: string;
  q?: HttpParams;
  foreignKey?: number;
  selfId?: number;

}